+++ .typedoc/custom-theme.mjs
@@ -19,6 +19,17 @@ class ClerkMarkdownTheme extends MarkdownTheme {
 }
 
 /**
+ * This map stores the comment for the first item in a union type.
+ * It'll be used to add that comment to the other items of the union type.
+ * This way the comment only has to be added once.
+ * @type {Map<string, import('typedoc').Comment>}
+ *
+ * The key is a concenation of the model's type name and the union type's declaration name.
+ * The value is the comment
+ */
+const unionCommentMap = new Map();
+
+/**
  * Our custom Clerk theme
  * @extends MarkdownThemeContext
  */
@@ -51,6 +62,63 @@ class ClerkMarkdownThemeContext extends MarkdownThemeContext {
         const output = superPartials.signature(customizedModel, customizedOptions);
 
         return output;
+      },
+      /**
+       * This condenses the output if only a "simple" return type + `@returns` is given.
+       * @param {import('typedoc').SignatureReflection} model
+       * @param {{ headingLevel: number }} options
+       */
+      signatureReturns: (model, options) => {
+        const defaultOutput = superPartials.signatureReturns(model, options);
+        const hasReturnsTag = model.comment?.getTag('@returns');
+
+        /**
+         * @type {any}
+         */
+        const type = model.type;
+        /**
+         * @type {import('typedoc').DeclarationReflection}
+         */
+        const typeDeclaration = type?.declaration;
+
+        /**
+         * Early return for non "simple" cases
+         */
+        if (!typeDeclaration?.signatures) {
+          if (model.type && this.helpers.hasUsefulTypeDetails(model.type)) {
+            return defaultOutput;
+          }
+        }
+        if (!hasReturnsTag) {
+          return defaultOutput;
+        }
+
+        /**
+         * Now the default output would be in this format:
+         *
+         * `Type`
+         *
+         * Contents of `@returns` tag
+         *
+         * It should be condensed to:
+         *
+         * `Type` — Contents of `@returns` tag
+         */
+
+        const o = defaultOutput.split('\n\n');
+
+        /**
+         * At this stage the output can be:
+         * - ['## Returns', '`Type`', 'Contents of `@returns` tag']
+         * - ['## Returns', '`Type`', '']
+         *
+         * We want to condense the first case by combining the second and third item with ` — `
+         */
+        if (o.length === 3 && o[2] !== '') {
+          return `${o[0]}\n\n${o[1]} — ${o[2]}`;
+        }
+
+        return defaultOutput;
       },
       /**
        * This hides the "Type parameters" section from the output
@@ -89,6 +157,30 @@ class ClerkMarkdownThemeContext extends MarkdownThemeContext {
         return output;
       },
       /**
+       * This ensures that everything is wrapped in a single codeblock
+       * @param {import('typedoc').DeclarationReflection} model
+       * @param {{ forceCollapse?: boolean }} [options]
+       */
+      declarationType: (model, options) => {
+        const defaultOutput = superPartials.declarationType(model, options);
+
+        // Ensure that the output starts with `\{ `. Some strings will do, some will have e.g. `\{<code>` or `\{[]`
+        const withCorrectWhitespaceAtStart = defaultOutput.startsWith('\\{ ')
+          ? defaultOutput
+          : defaultOutput.startsWith('\\{')
+            ? defaultOutput.replace('\\{', '\\{ ')
+            : defaultOutput;
+
+        const output = withCorrectWhitespaceAtStart
+          // Remove any backticks
+          .replace(/`/g, '')
+          // Remove any `<code>` and `</code>` tags
+          .replace(/<code>/g, '')
+          .replace(/<\/code>/g, '');
+
+        return `<code>${output}</code>`;
+      },
+      /**
        * This modifies the output of union types by wrapping everything in a single `<code>foo | bar</code>` tag instead of doing `<code>foo</code>` | `<code>bar</code>`
        * @param {import('typedoc').UnionType} model
        */
@@ -107,7 +199,7 @@ class ClerkMarkdownThemeContext extends MarkdownThemeContext {
         return `<code>${output}</code>`;
       },
       /**
-       * This ensures that everything is wrapped in a single codeblock (not individual ones)
+       * This ensures that everything is wrapped in a single codeblock
        * @param {import('typedoc').SignatureReflection[]} model
        * @param {{ forceParameterType?: boolean; typeSeparator?: string }} [options]
        */
@@ -132,7 +224,7 @@ class ClerkMarkdownThemeContext extends MarkdownThemeContext {
       /**
        * Copied from original theme.
        * Changes:
-       * - Remove summaries over tables (TODO: Use elementSummaries instead)
+       * - Remove summaries over tables and instead use `@unionReturnHeadings` to add headings
        * - Only use one newline between items
        * https://github.com/typedoc2md/typedoc-plugin-markdown/blob/7032ebd3679aead224cf23bffd0f3fb98443d16e/packages/typedoc-plugin-markdown/src/theme/context/partials/member.typeDeclarationUnionContainer.ts
        * @param {import('typedoc').DeclarationReflection} model
@@ -143,10 +235,51 @@ class ClerkMarkdownThemeContext extends MarkdownThemeContext {
          * @type {string[]}
          */
         const md = [];
+        /**
+         * @type {string[]}
+         */
+        const headings = [];
+        /**
+         * Search for the `@unionReturnHeadings` tag in the comment of the model
+         */
+        const unionReturnHeadings = model.comment?.getTag('@unionReturnHeadings');
+
         if (model.type instanceof UnionType) {
           const elementSummaries = model.type?.elementSummaries;
           model.type.types.forEach((type, i) => {
             if (type instanceof ReflectionType) {
+              if (unionReturnHeadings && unionReturnHeadings.content.length > 0) {
+                const content = this.helpers.getCommentParts(unionReturnHeadings.content);
+                const unionHeadings = JSON.parse(content);
+
+                /**
+                 * While iterating over the union types, the headings are pulled from `@unionReturnHeadings` and added to the array
+                 */
+                headings.push(unionHeadings[i]);
+
+                /**
+                 * The `model.type.types` is the array of the individual items of the union type.
+                 * We're documenting our code by only adding the comment to the first item of the union type.
+                 *
+                 * In this block, we're doing the following:
+                 * 1. Generate an ID for the item in the unionCommentMap
+                 * 2. Check if the union type has a comment (truthy for the first item)
+                 * 3. Add the comment to the map
+                 * 4. If the union doesn't have a comment for the given ID, add the comment from the map to the item
+                 */
+                if (type.declaration.children) {
+                  for (const decl of type.declaration.children) {
+                    const id = `${model.name}-${decl.name}`;
+
+                    if (decl.comment && !unionCommentMap.has(id)) {
+                      unionCommentMap.set(id, decl.comment);
+                    } else if (!decl.comment && unionCommentMap.has(id)) {
+                      decl.comment = unionCommentMap.get(id);
+                    }
+                  }
+                }
+              }
+
               md.push(this.partials.typeDeclarationContainer(model, type.declaration, options));
             } else {
               md.push(`${this.partials.someType(type)}`);
@@ -156,7 +289,33 @@ class ClerkMarkdownThemeContext extends MarkdownThemeContext {
             }
           });
         }
-        return md.join('\n');
+
+        if (!unionReturnHeadings) {
+          return md.join('\n');
+        }
+
+        const items = headings.map(i => `'${i}'`).join(', ');
+        const tabs = md.map(i => `<Tab>${i}</Tab>`).join('\n');
+
+        return `<Tabs items={[${items}]}>
+${tabs}
+</Tabs>`;
+      },
+      /**
+       * This ensures that everything is wrapped in a single codeblock
+       * @param {import('typedoc').ArrayType} model
+       */
+      arrayType: model => {
+        const defaultOutput = superPartials.arrayType(model);
+
+        const output = defaultOutput
+          // Remove any backticks
+          .replace(/`/g, '')
+          // Remove any `<code>` and `</code>` tags
+          .replace(/<code>/g, '')
+          .replace(/<\/code>/g, '');
+
+        return `<code>${output}</code>`;
       },
     };
   }