+++ packages/backend/src/api/resources/index.ts
@@ -1,8 +1,12 @@
+export * from './ActorToken';
 export * from './AccountlessApplication';
 export * from './AllowlistIdentifier';
+export * from './BlocklistIdentifier';
 export * from './Client';
+export * from './CnameTarget';
 export * from './Cookies';
 export * from './DeletedObject';
+export * from './Domain';
 export * from './Email';
 export * from './EmailAddress';
 
@@ -23,7 +27,9 @@ export * from './InstanceRestrictions';
 export * from './InstanceSettings';
 export * from './Invitation';
 export * from './JSON';
+export * from './JwtTemplate';
 export * from './OauthAccessToken';
+export * from './OAuthApplication';
 export * from './Organization';
 export * from './OrganizationInvitation';
 export * from './OrganizationMembership';