+++ packages/clerk-js/src/utils/componentGuards.ts
@@ -22,6 +22,6 @@ export const disabledOrganizationsFeature: ComponentGuard = (_, environment) => 
   return !environment?.organizationSettings.enabled;
 };
 
-export const disabledCommerceFeature: ComponentGuard = (_, environment) => {
-  return !environment?.__experimental_commerceSettings.enabled;
+export const disabledBillingFeature: ComponentGuard = (_, environment) => {
+  return !environment?.__experimental_commerceSettings.billing.enabled;
 };