+++ packages/shared/src/react/types.ts
@@ -89,18 +89,21 @@ export type PaginatedResourcesWithDefault<T> = {
  */
 export type PaginatedHookConfig<T> = T & {
   /**
-   * If `true`, newly fetched data will be appended to the existing list rather than replacing it. Useful for implementing infinite scroll functionality. Defaults to `false`.
+   * If `true`, newly fetched data will be appended to the existing list rather than replacing it. Useful for implementing infinite scroll functionality.
+   * @default false
    */
   infinite?: boolean;
   /**
-   * If `true`, the previous data will be kept in the cache until new data is fetched. Defaults to `false`.
+   * If `true`, the previous data will be kept in the cache until new data is fetched.
+   * @default false
    */
   keepPreviousData?: boolean;
 };
 
 export type PagesOrInfiniteConfig = PaginatedHookConfig<{
   /**
-   * If `true`, a request will be triggered. Defaults to `true`.
+   * If `true`, a request will be triggered.
+   * @default true
    */
   enabled?: boolean;
 }>;
@@ -110,11 +113,13 @@ export type PagesOrInfiniteConfig = PaginatedHookConfig<{
  */
 export type PagesOrInfiniteOptions = {
   /**
-   * A number that specifies which page to fetch. For example, if `initialPage` is set to 10, it will skip the first 9 pages and fetch the 10th page. Defaults to `1`.
+   * A number that specifies which page to fetch. For example, if `initialPage` is set to 10, it will skip the first 9 pages and fetch the 10th page.
+   * @default 1
    */
   initialPage?: number;
   /**
-   * A number that specifies the maximum number of results to return per page. Defaults to `10`.
+   * A number that specifies the maximum number of results to return per page.
+   * @default 10
    */
   pageSize?: number;
 };