+++ packages/types/src/commerceSettings.ts
@@ -4,12 +4,17 @@ import type { ClerkResourceJSON } from './json';
 import type { ClerkResource } from './resource';
 
 export interface __experimental_CommerceSettingsJSON extends ClerkResourceJSON {
-  enabled: boolean;
-  stripe_publishable_key: string;
+  billing: {
+    enabled: boolean;
+    stripe_publishable_key: string;
+  };
 }
 
 export interface __experimental_CommerceSettingsResource extends ClerkResource {
-  enabled: boolean;
-  stripePublishableKey: string;
+  billing: {
+    enabled: boolean;
+    stripePublishableKey: string;
+  };
+
   __internal_toSnapshot: () => __experimental_CommerceSettingsJSONSnapshot;
 }