+++ packages/types/src/elementIds.ts
@@ -52,4 +52,4 @@ export type OrganizationPreviewId =
 export type CardActionId = 'havingTrouble' | 'alternativeMethods' | 'signUp' | 'signIn' | 'usePasskey' | 'waitlist';
 
 export type MenuId = 'invitation' | 'member' | ProfileSectionId;
-export type SelectId = 'countryCode' | 'role';
+export type SelectId = 'countryCode' | 'role' | 'paymentSource';